requirements.txt
asgiref==3.3.4  
autopep8==1.5.7  
Django==3.2.4 
django-filter==2.4.0 
pip==21.1.3
pycodestyle==2.7.0  
pytz==2021.1 
setuptools==41.2.0 
sqlparse==0.4.1  
toml==0.10.2 